+++ packages/component/src/types/external/OrgSchema/Thing.spec.ts
