+++ packages/component/src/types/external/OrgSchema/Thing.ts
