+++ packages/component/src/types/external/OrgSchema/VoteAction.spec.ts
