+++ packages/component/src/types/external/OrgSchema/VoteAction.ts
